+++ Changelog.md
@@ -2,10 +2,18 @@
 ## Breaking Changes
 ### JLib.Reflection
 - removed all typePackage usages and implementations. Use the TypePackageBuilder instead.
+  - calls will be cached by default. 
+
+### JLib.Helper
+- changed ReflectionHelper attribute method declarations to cache the result.
 
 ### JLib.ValueTypes.*
 - renamed JLib.ValueTypes.Mapping -> JLib.ValueTypes.AutoMapper
 - moved valueType json converters to JLib.ValueTypes package
+
+## Features
+### JLib.Helper
+- AttributeCache added
 
 # 0.12.0
 ## Breaking Changes